requirements.txt
fastapi==0.116.0
python-dotenv==1.1.1
groq==0.30.0
sentence-transformers==5.0.0
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1